Refactor into different files
