+++ send_email_messages/frontend/index.js
@@ -1,70 +1,19 @@
-import {initializeBlock, Box, Button, useRecords, useBase} from '@airtable/blocks/ui';
-import {base} from '@airtable/blocks';
-import React, {useState} from 'react';
-
+import { initializeBlock, Box, Button, useRecords, useBase } from '@airtable/blocks/ui';
+import { base } from '@airtable/blocks';
+import React, { useState } from 'react';
+import { getConfigOfType, getGlobalValueFromKey, getEmailTypes } from './config.js';
+import { SendMessagesStep } from './send-messages.js'
 
 function findById(table, id) {
     return table.find(r => r.id === id);
 }
 
-async function getConfigOfType(type) {
-	const allConfig = (await base.getTable("Email Configuration").selectRecordsAsync()).records;
-	return allConfig.filter(c => c.getCellValue("type").name === type)
-}
-
-async function getGlobalValueFromKey(key) {
-    const globalConfig = await getConfigOfType("global");
-    return globalConfig.find(k => k.name === key).getCellValue("value");
-}
-
-async function getSendgridConfig() {
-	const globalConfig = await getConfigOfType("global");
-
-	function globalValueFromKey(key) {
-        return globalConfig.find(k => k.name === key).getCellValue("value");
-	}
-
-	return {
-		SENDGRID_PROXY_TOKEN: globalValueFromKey("SENDGRID_PROXY_TOKEN"),
-		REPLY_TO: {
-    	    name: globalValueFromKey("reply_to_name"),
-    	    email: globalValueFromKey("reply_to_email")
-		},
-		EMAIL_TYPES: await getEmailTypes()
-	}
-}
-
-async function getEmailTypes() {
-	const emailTypesConfig = await getConfigOfType("email-template");
-
-	const EMAIL_TYPES = {};
-	for (const emailType of emailTypesConfig) {
-	    EMAIL_TYPES[emailType.name] = {
-	        "sendgrid_template": emailType.getCellValue("sendgrid_template"),
-	        "audience": emailType.getCellValue("audience").name,
-            "stages": emailType.getCellValue("stages").map((s) => {return s.name}),
-	    }
-	}
-	return EMAIL_TYPES;
-}
-
 // Customize what fields we pull in from a delivery for an email template.
-async function getEmailInfoForDelivery(delivery) {
+async function getEmailInfoForDelivery(delivery, volunteer) {
     let output = {
         delivery: {},
         volunteer: {}
     }
-    
-    // see if there's a volunteer assigned
-    // TODO inefficient do not load this volunteers this many times!! do this in the outer function
-   	const allVolunteers = (await base.getTable("Volunteers").selectRecordsAsync()).records;
-
-    let volunteer;
-    const volunteers = delivery.getCellValue("Assigned volunteer");
-    if (volunteers && volunteers.length > 0) {
-        volunteer = findById(allVolunteers, volunteers[0].id);
-    }
-    
     const templateVarsConfig = await getConfigOfType("template-variable");
 
     for (let r of templateVarsConfig) {
@@ -82,8 +31,8 @@ async function getEmailInfoForDelivery(delivery) {
     return output;
 }
 
-async function formatSengridMessage(delivery, messageType, audience) {
-    const sendgridData = await getEmailInfoForDelivery(delivery);
+async function formatMessageRecord(messageType, delivery, volunteer, audience) {
+    const sendgridData = await getEmailInfoForDelivery(delivery, volunteer);
 
     var toEmail;
     if (audience === "volunteer") {
@@ -117,6 +66,8 @@ async function computeMessagesToCreate(addWarning) {
     }
 	const allDeliveries = (await base.getTable("Deliveries").selectRecordsAsync()).records;
 	const allMessages = (await base.getTable("Messages").selectRecordsAsync()).records;
+    const allVolunteers = (await base.getTable("Volunteers").selectRecordsAsync()).records;
+
 	const EMAIL_TYPES = await getEmailTypes();
 
     for (const delivery of allDeliveries) {
@@ -133,6 +84,13 @@ async function computeMessagesToCreate(addWarning) {
             existingMessageTypes = [];
         }
 
+        // get the assigned volunteer, if any
+        let volunteer;
+        const volunteers = delivery.getCellValue("Assigned volunteer");
+        if (volunteers && volunteers.length > 0) {
+            volunteer = findById(allVolunteers, volunteers[0].id);
+        }
+
         for (let messageType of Object.keys(EMAIL_TYPES)) {
             if (existingMessageTypes.includes(messageType)) {
                 // We've already sent a message of this type for this delivery
@@ -143,7 +101,7 @@ async function computeMessagesToCreate(addWarning) {
                 continue;
             }
 
-            const {message, warning} = await formatSengridMessage(delivery, messageType, EMAIL_TYPES[messageType]["audience"]);
+            const {message, warning} = await formatMessageRecord(messageType, delivery, volunteer, EMAIL_TYPES[messageType]["audience"]);
             if (message) {
                 output.messagesToCreate.push(message);
             }
@@ -153,86 +111,6 @@ async function computeMessagesToCreate(addWarning) {
         }
     }
     return output;
-}
-
-async function sendMessage(messageToSend, cfg) {
-    const dynamicTemplateData = JSON.parse(messageToSend.getCellValue("Template Data"));
-    const sendgridData = {
-        "personalizations": [
-            {
-                "to": [
-                    {
-                    "email": messageToSend.getCellValue("Recipient")
-                    }
-                ],
-                "dynamic_template_data": dynamicTemplateData
-            }
-        ],
-        "from": {
-            "name": "Neighbor Express",
-            "email": "noreply@neighborexpress.org"
-        },
-        "reply_to": cfg.REPLY_TO,
-        "template_id": cfg.EMAIL_TYPES[messageToSend.getCellValue("Email type").name]["sendgrid_template"]
-    }
-
-    const response = await fetch(`https://nex-sendgrid-proxy.geoffreylitt.now.sh/api/sendgrid-proxy?proxy_token=${cfg.SENDGRID_PROXY_TOKEN}`, {
-        method: 'POST',
-        body: JSON.stringify(sendgridData)
-    });
-
-    if (response.status === 202) {
-        base.getTable("Messages").updateRecordAsync(messageToSend.id, {
-            "Status": { name: "Sent" },
-            "Sent Time": new Date()
-        })
-    } else {
-        base.getTable("Messages").updateRecordAsync(messageToSend.id, {
-            "Status": { name: "Errored" },
-        })
-    }
-}
-
-
-function StepSendMessages() {
-	const [result, setResult] = useState(undefined);
-
-	const messagesToSend = useRecords(useBase().getTable("Messages")).filter(m => m.getCellValue("Status").name === "Queued");
-
-	function cancelButton() {
-		// also disable all buttons!
-		setResult('**Canceled!** No messages sent.');
-	}
-
-	async function sendMessages() {
-		const cfg = await getSendgridConfig();
-        for (const messageToSend of messagesToSend) {
-            await sendMessage(messageToSend, cfg);
-        }
-		setResult('**Done sending messages!**');
-	}
-
-	return (
-		<Box>
-			<h2> Step 2: Send emails </h2>
-            {result ? <p> {result} </p> : <>
-        	  	{
-        	  		messagesToSend.length === 0 ?
-        	  		<p> No enqueued messages to send. <strong> You're all done! </strong></p> :
-        	  		<>
-        	  			<p> {messagesToSend.length} messages queued to send. Review the list in the Messages tab, then click below to send them. </p>
-        	  			<Button variant="secondary" onClick={cancelButton} disabled={result!=undefined}>
-        			  		Cancel
-        			  	</Button>
-        	  			<Button variant="primary" onClick={sendMessages} disabled={result!=undefined}> 
-        	  				Send {messagesToSend.length} Messages
-        			  	</Button>
-        	  		</>
-        	  	}
-            </>}
-    	  	
-        </Box>
-	)
 }
 
 function WarningAccordion({warnings}) {
@@ -264,6 +142,7 @@ function MainUIBlock() {
     }
 
     async function refreshQueue() {
+        console.log("hello");
     	var result = ''
 
         // Clear out the warnings from any previous run
@@ -293,7 +172,7 @@ function MainUIBlock() {
             <WarningAccordion warnings={warnings}/>
             {step1result && <>
                 <p> {step1result} </p>
-                <StepSendMessages/>
+                <SendMessagesStep/>
             </>}
         </Box>
     );
